+++ app/src/main/java/com/example/moviereview/MainActivity.kt
@@ -5,6 +5,8 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.moviereview.databinding.ActivityMainBinding
+import com.example.moviereview.review.ReviewFragment
+import com.example.moviereview.search.SearchFragment
 import com.google.android.material.tabs.TabLayoutMediator
 
 class MainActivity : AppCompatActivity() {
