+++ app/src/main/java/com/example/moviereview/room/ReviewDAO.kt
@@ -1,9 +1,8 @@
-package com.example.moviereview
+package com.example.moviereview.room
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
-import androidx.room.Update
 import com.example.moviereview.network.MovieResponse
 
 @Dao
