+++ app/src/main/java/com/example/moviereview/room/ReviewDB.kt
@@ -1,7 +1,8 @@
-package com.example.moviereview
+package com.example.moviereview.room
 
 import android.content.Context
 import androidx.room.*
+import com.example.moviereview.MovieItemTypeConverter
 import com.google.gson.Gson
 
 @Database(entities = [ReviewModel::class], version = 1)
