+++ app/src/main/java/com/example/moviereview/search/SearchAdapter.kt
@@ -1,12 +1,10 @@
-package com.example.moviereview
+package com.example.moviereview.search
 
 import android.os.Build
 import android.text.Html
 import android.util.Log
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
