+++ app/src/main/java/com/example/moviereview/search/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moviereview
+package com.example.moviereview.search
 
 import android.content.Intent
 import android.net.Uri
@@ -9,6 +9,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.lifecycle.MutableLiveData
+import com.example.moviereview.R
+import com.example.moviereview.review.ReviewDialog
 import com.example.moviereview.databinding.FragmentSearchBinding
 import com.example.moviereview.network.MovieResponse
 import com.example.moviereview.network.RetrofitService
